Add MessageTypeCacheBuilder
This is a helper class to simplify dependency injection configuration.

We need sometimes to both register dependencies in IServiceCollection as
well as register custom serialization types for serializer to know how
to map them. Hence we are combining these two concerns in this
dependency builder.

Also libraries such as Protobuf that depend on static runtime classes to
register types - can add custom Action<IMessageTypeCache> post-build
actions that will be executed in the order of registration after the
initial build of MessageTypeCache and before it is returned to the
requested party.

The IMessageTypeCache is registered as singleton (hardcoded, not
changeable). However, in future it's possible to include the flag on
MessageTypeCacheBuilder indicating whether to build it as singleton or
no.