Add MessageHandlerFactory that uses IServiceProvider
This message handler factory is a service locator that uses .NET Core DI
container to resolve instances of IMessageHandler<T>. It may be
anti-pattern and it may not work so well in ASP.NET Core scenario (with
scoped services, since the DI will be disposed after Startup has been
build). We need to test it and to come up with better solution if any.