Registe IMessageHandlerFactory as singleton for SignalR
When we use ASP.Net core's Startup.Configure to register our services,
the parent container gets disposed and we need to capture it by
registering MessageHandlerFactory as singleton.

This might not be the best solution (maybe we can use some kind of
scoped services approach), but for now it works.