Add Broadcast message type and handler
It can be used for events that simply need to be broadcasted to all the
clients in the same messaging group. For example, starting to attacking
a player need to be drawn on the front-end side so all the players see
who is attacking who. This event will be sent from client to the server
and the server will re-send it to all the other clients for them to see
what's happening.