Add TCP server for RopeWar and refactor Authentication
Refactor Authentication so that it can be enabled/disabled and it can
use either Auth0 provider or local (offline/stub) provider.
Authentication now can be constructed using TyrAuthenticationBuilder.

TCP implementation of RopeWar domain will be the only TCP implementation
for now (movement domain will be possibly removed later on). It will
remain here and will be updated to be a 'reference' tcp server so that
we keep implementations abstract from web/signalr technologies.