Use profile API client (ICharactersClient) everywhere
Use this statically-typed API client in Authentication and RopeWar
projects. Now we don't need adapters there, but they are left there for
now to facilitate SOLID / dependency inversion. However, for now this is
additional abstraction level that doesn't improve anything and so it can
be omitted later on to make the code more readable.