Add acknowledgement to messages
If message is of Message type and has MessageId field set, the
AcknowledgeReceived message will be sent back after the message has been
received by server.

This functionality is need for multi-message flows, such as
authentication flow where we need to first send Authenticate message,
and then Connect message, in this exact order.