Always put JsonSerializedMessage as first message in message type cache
When client has different number of registered messages than the server,
messages are being sorted by name so discrepancy occurs where some
messages on the client side might not represent the same messages on the
server side, and everything goes to hell.

The temporary hacky solution to this is to use Json serialization over
protobuf so that we only need to have a single message registered in
protobuf system, and json serialization/deserialization will take place
based on message "string type identifier" representation (e.g. full type
.NET name, as of now).

However, the case exists when some messages are registered on the client
that go before JsonSerializationMessage alphabetically (e.g. Abort), but
on the server side it's not registered, leading to
JsonSerilalizationMessage having different identifiers on client and
server sides. To counter this problem, we hardcode the solution that if
this message is registered at all (present in message type cache) - we
put it into the first place, so the identity of this message will always
be "1" (in protobuf).