Send metadata with messages, split network and serialization layers
This commit contains a couple of important changes to the overall
architecture of handling messages.

Now we have a clear and distinct two levels of abstraction layers:

- Network layer (sending data over the network)
- Serialization layer (serializing the data and deserializing the
  incoming data)

Another important change is: now we can send Metadata together with
messages from client to server.

Client and server abstractions are separated using separate interfaces
and classes so that now we can develop them independently (they are in
the same assemblies for now).

Overall, these changes are handled by this single commit:

- Changed approach: serializating and network layer are separated
-- As a result, now TCP connection can use JSON serialization, and
SignalR (json) connection can use Protobuf for message data
serialization / deserialization. This is how hosts and test client are
configured right now.
- Provided a way to send MessageMetadata: it can either be generated
  (for Idempotency implementation later on) or sent by the user using
  IConnection API.
- Split client and server code so it's easier to understand what is
  going on.

P.S. 22 tests are failing now. They need to be rewritten / fixed.