Implement proper reconnection with re-authentication
- Use HttpClientFactory to create character in TestClient

- Authenticate client during reconnection

We need to send Authenticate & Connect messages again during reconnection, do
this by saving CharacterId inside MessageProcessor when initially connecting.

- Use ConcurrentDictionary for subscribing (thread-safe)

If we don't do this - we have an issue that when we are dispatching messages
to registered handlers and at the same time somebody subscribes - we've got
enumeration modified exception because it's not thread safe.
ConcurrentDictionary's Values property getter is thread safe and returns the
collection of values at the time when requested.

- Do not lock for waiting for reconnect to finish in Send methods

If we lock inside Send methods and wait for reconnection to finish, we create
a deadlock, because reconnection itself calls Send methods (to authenticate
after reconnection). We need to come up with other way of postponing Send
actions until reconnection is done.

- Disconnect the SignalR client from the server side if exception happens or
client is disconnected (use context.Abort())

When exception happens client is removed from connected clients list but the
SignalR connection is not destroyed. This commit fixes this.

Client can learn about disconnected connection only by doing constant pings so
later on when we have pings functionality - when ping won't be successful the
client would reconnect automatically.