Re-implement acknowledgement using TaskCompletionSource
If we don't do Task.Yield() after awaiting for TCS, it doesn't work as
expected: the message sent from the server to the client doesn't arrive
to the client until the next message is being sent.

This is very weird behaviour that needs further investigation. For now
I'm leaving Task.Yield() in code.