Acknowledge handled messages
Introduce AcknowledgeHandled message that is being sent from the server
to the client once the message has been successfully handled, if the
client specified this type of acknowledgement.

Refactor acknowledgement to have multiple types (enum) instead of a bool
flag.

This is enabled by default so now everything is slow and highly
synchronous. Optimizations might be done later on for specific messages
that can behave like fire-and-forget, but for now we need this kind of
stability.