Add ConnectionFactory factory to be able to connect to multiple servers
Now we can specify connection string during connection time so that we
can have multiple MessageProcessor instances.

Maybe it's not the best way to implement things using FactoryFactory, we
need to think about refactoring this code. This was done like this for
backward compatibility so that all the old code would work as it works
now.