Prepare for entering rope-war in Characters API
Whenever voting to start ropewar has been completed - we send API
call(s) to Characters API to set the status of given characters to
RopeWar activity so that World actions would be blocked and RopeWar
actions would be allowed.

This commit introduces basic domain events concept: we would have list
of events on the aggregate root that will be dispatched during save
process (and possibly stored in the DB later in more advanced
implementation for compensatory actions).

But for now we cannot send the request to Characters API service because
LocationStore doesn't know anything about ConnectedClientContext and
this is what is being injected together with IHttpClient if we inject it
into LocationStore which is a singleton, but Context is scoped instance
(obviously).

We need a way to send the request using some kind of Server-Side
IHttpClient singleton instance and not a service-scoped one. Most likely
by introducing a separate interface for such concerns.