Introduce scoped LocationRepository
We needed to have Scoped dependency for LocationStore because it needs
IHttpClient which depends on client Context which is a scoped dependency
and cannot be injected into singleton, however we still need all the
Locations data to be Singleton (it's an in-memory store). For this, we
split LocationStore to LocationStore and LocationRepository, where
LocationStore holds the data of all locations and is Singleton, and
LocationRepository works over the LocationStore and is Scoped dependency
so that IHttpClient can be injected into LocationRepository instance.