Do not fail validation even if there are some trailing characters
We cannot fail validation as soon as there are some trailing events
because user might must needs to correct his input before submitting it.

We need to find a way to better validate this so there are no trailing
characters after complete text submit, hence the code is still there as
a reminder of work that needs to be done.