Set PrivateAssets for Project references
During development, I am setting also PackageReferences to
PrivateAssets, but during build time they should be disabled.

For now I'm doing it manually cause I don't know how to set it
conditional to the build so that in initial build PrivateAssets are
there, but when you actually run the project - they are not. Probably
it's not possible to do what I want, so I'm continuing to do this
manually.